Move JSF 2.3 Spec Issue tests to their own test class
