+++ dev/com.ibm.ws.jsf.2.3_fat/fat/src/com/ibm/ws/jsf23/fat/FATSuite.java
@@ -35,6 +35,7 @@ import com.ibm.ws.jsf23.fat.tests.JSF23IterableSupportTests;
 import com.ibm.ws.jsf23.fat.tests.JSF23JPA22Test;
 import com.ibm.ws.jsf23.fat.tests.JSF23MapSupportTests;
 import com.ibm.ws.jsf23.fat.tests.JSF23SelectOneRadioGroupTests;
+import com.ibm.ws.jsf23.fat.tests.JSF23SpecIssueTests;
 import com.ibm.ws.jsf23.fat.tests.JSF23UIRepeatConditionTests;
 import com.ibm.ws.jsf23.fat.tests.JSF23UISelectManyTests;
 import com.ibm.ws.jsf23.fat.tests.JSF23ViewParametersTests;
@@ -88,7 +89,8 @@ import componenttest.topology.impl.JavaInfo;
                 JSF23CDIFacesInMetaInfTests.class,
                 JSF23CDIFacesInWebXMLTests.class,
                 JSF23CDIConfigByACPTests.class,
-                Faces30Tests.class
+                Faces30Tests.class,
+                JSF23SpecIssueTests.class
 })
 
 public class FATSuite {
