+++ dev/com.ibm.ws.jsf.2.3_fat/publish/servers/jsf23SpecIssueServer/.gitignore
@@ -1,0 +1,1 @@
+/dropins
