+++ dev/com.ibm.ws.jsf.2.3_fat/publish/servers/jsf23SpecIssueServer/server.xml
@@ -1,0 +1,18 @@
+<!--
+    Copyright (c) 2022 IBM Corporation and others.
+    All rights reserved. This program and the accompanying materials
+    are made available under the terms of the Eclipse Public License v1.0
+    which accompanies this distribution, and is available at
+    http://www.eclipse.org/legal/epl-v10.html
+   
+    Contributors:
+        IBM Corporation - initial API and implementation
+ -->
+<server description="Server for testing JavaServer Faces 2.3 spec issues">
+    <include location="../fatTestPorts.xml"/>
+
+    <featureManager>
+        <feature>jsf-2.3</feature>
+        <feature>cdi-2.0</feature>
+    </featureManager>
+</server>
